[N/A] Changes to make life easier when generating types
